changes : updated file structure and error handling
