+++ package.json
@@ -2,10 +2,56 @@
   "name": "solone-sdk",
   "version": "1.0.6",
   "description": "javascript sdk for solona",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "check-types": "tsc --noEmit",
+    "clean": "rimraf build",
+    "build:js": "tsc",
+    "build": "npm-run-all clean build:js",
+    "test": "jest",
+    "test:clear-cache": "jest --clearCache",
+    "test:coverage": "jest --coverage",
+    "coverage": "npm-run-all test:*",
+    "lint": "eslint 'src/**/*.ts' --fix",
+    "docs:build": "typedoc",
+    "docs:serve": "serve docs",
+    "serve-docs": "npm-run-all docs:*",
+    "pre-commit": "npm-run-all check-types lint"
   },
+  "author": "aqadir64",
+  "license": "ISC",
+  "dependencies": {
+    "@solana/web3.js": "^1.29.0",
+    "borsh": "^0.6.0",
+    "typescript": "^4.4.3"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.32.0",
+    "@typescript-eslint/parser": "^4.32.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^7.0.2",
+    "prettier": "^2.4.1",
+    "@types/jest": "^26.0.19",
+    "jest": "^26.6.3",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "^3.0.2",
+    "serve": "^11.3.2",
+    "ts-jest": "^26.4.4",
+    "ts-node": "^9.1.1",
+    "typedoc": "^0.20.5",
+    "typescript": "^4.1.3"
+  },
+  "files": [
+    "dist/**/*"
+  ],
+  "repository": "https://github.com/solone-io/Solone-SDK.git",
+  "bugs": {
+    "url": "https://github.com/solone-io/Solone-SDK/issues"
+  },
+  "homepage": "https://github.com/solone-io/Solone-SDK#readme",
   "keywords": [
     "solana",
     "solana-sdk",
@@ -19,21 +65,5 @@
     "solana-development",
     "api",
     "@solana/web3.js"
-  ],
-  "author": "aqadir64",
-  "license": "ISC",
-  "dependencies": {
-    "@solana/web3.js": "^1.29.0",
-    "borsh": "^0.6.0",
-    "typescript": "^4.4.3"
-  },
-  "devDependencies": {},
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/solone-io/Solone-SDK.git"
-  },
-  "bugs": {
-    "url": "https://github.com/solone-io/Solone-SDK/issues"
-  },
-  "homepage": "https://github.com/solone-io/Solone-SDK#readme"
+  ]
 }
